refactor: cache mutations
