+++ packages/postgrest-react-query/src/mutate/use-delete-mutation.ts
@@ -39,10 +39,10 @@ function useDeleteMutation<
 ) {
   const queriesForTable = useQueriesForTableLoader(getTable(qb));
   const deleteItem = useDeleteItem({
+    ...opts,
     primaryKeys,
     table: getTable(qb),
     schema: qb.schema as string,
-    opts,
   });
 
   return useMutation({
