+++ packages/postgrest-react-query/src/mutate/use-insert-mutation.ts
@@ -37,10 +37,10 @@ function useInsertMutation<
 ) {
   const queriesForTable = useQueriesForTableLoader(getTable(qb));
   const upsertItem = useUpsertItem({
+    ...opts,
     primaryKeys,
     table: getTable(qb),
     schema: qb.schema as string,
-    opts,
   });
 
   return useMutation({
