+++ packages/postgrest-react-query/src/mutate/use-upsert-mutation.ts
@@ -37,10 +37,10 @@ function useUpsertMutation<
 ) {
   const queriesForTable = useQueriesForTableLoader(getTable(qb));
   const upsertItem = useUpsertItem({
+    ...opts,
     primaryKeys,
     table: getTable(qb),
     schema: qb.schema as string,
-    opts,
   });
 
   return useMutation({
