+++ packages/postgrest-swr/src/cache/use-upsert-item.ts
@@ -1,6 +1,6 @@
 import {
   upsertItem,
-  UpsertItemProps,
+  UpsertItemOperation,
 } from '@supabase-cache-helpers/postgrest-core';
 import { MutatorOptions, useSWRConfig } from 'swr';
 
@@ -16,13 +16,13 @@ import { getMutableKeys } from '../lib/mutable-keys';
  * @returns A function that takes a record of type `Type` and returns a promise that resolves once the record has been upserted into the cache.
  * **/
 export function useUpsertItem<Type extends Record<string, unknown>>(
-  opts: Omit<UpsertItemProps<Type>, 'input'> & MutatorOptions<Type>,
+  opts: Omit<UpsertItemOperation<Type>, 'input'> & MutatorOptions<Type>,
 ): (input: Type) => Promise<void> {
   const { mutate, cache } = useSWRConfig();
   const getPostgrestFilter = usePostgrestFilterCache();
 
   return async (input: Type) =>
-    await upsertItem(
+    await upsertItem<string, Type>(
       {
         input,
         ...opts,
@@ -32,7 +32,7 @@ export function useUpsertItem<Type extends Record<string, unknown>>(
         getPostgrestFilter,
         mutate: (key, data) => {
           mutate(key, data, {
-            ...opts.opts,
+            ...opts,
             revalidate: opts?.revalidate ?? false,
           });
         },
