+++ packages/postgrest-swr/src/mutate/use-insert-mutation.ts
@@ -40,11 +40,10 @@ function useInsertMutation<
   const key = useRandomKey();
   const queriesForTable = useQueriesForTableLoader(getTable(qb));
   const upsertItem = useUpsertItem({
+    ...opts,
     primaryKeys,
     table: getTable(qb),
     schema: qb.schema as string,
-    opts,
-    ...opts,
   });
 
   return useMutation<R[] | null, PostgrestError, string, T['Insert'][]>(
