+++ packages/postgrest-swr/src/mutate/use-upsert-mutation.ts
@@ -40,11 +40,10 @@ function useUpsertMutation<
   const key = useRandomKey();
   const queriesForTable = useQueriesForTableLoader(getTable(qb));
   const upsertItem = useUpsertItem({
+    ...opts,
     primaryKeys,
     table: getTable(qb),
     schema: qb.schema as string,
-    opts,
-    ...opts,
   });
 
   return useMutation<R[] | null, PostgrestError, string, T['Insert'][]>(
