+++ packages/postgrest-swr/src/subscribe/use-subscription.ts
@@ -6,7 +6,7 @@ import {
   REALTIME_POSTGRES_CHANGES_LISTEN_EVENT,
   SupabaseClient,
 } from '@supabase/supabase-js';
-import { PostgrestMutatorOpts } from '@supabase-cache-helpers/postgrest-core';
+import { RevalidateOpts } from '@supabase-cache-helpers/postgrest-core';
 import { useEffect, useState } from 'react';
 import { MutatorOptions as SWRMutatorOptions } from 'swr';
 
@@ -15,7 +15,7 @@ import { useDeleteItem, useUpsertItem } from '../cache';
 /**
  * Options for `useSubscription` hook.
  */
-export type UseSubscriptionOpts<T extends GenericTable> = PostgrestMutatorOpts<
+export type UseSubscriptionOpts<T extends GenericTable> = RevalidateOpts<
   T['Row']
 > &
   SWRMutatorOptions & {
@@ -51,18 +51,16 @@ function useSubscription<T extends GenericTable>(
 ) {
   const [status, setStatus] = useState<string>();
   const deleteItem = useDeleteItem({
+    ...opts,
     primaryKeys,
     table: filter.table,
     schema: filter.schema,
-    opts,
-    ...opts,
   });
   const upsertItem = useUpsertItem({
+    ...opts,
     primaryKeys,
     table: filter.table,
     schema: filter.schema,
-    opts,
-    ...opts,
   });
 
   useEffect(() => {
